+++ src/dtactions/uniactions/actionclasses/komodo-actions.py
@@ -8,7 +8,7 @@ import time
 
 import natlink
 # from dtactions import messagefunctions as mf
-from dtactions.unimacroactionclasses.actionbases import AllActions
+from dtactions.actionclasses.actionbases import AllActions
 # from dtactions.unimacro.unimacroactions import doAction as action
 from dtactions.sendkeys import sendkeys as keystroke
